Initial Commit
CS 4381/5381: TSE -- Mobile Game Programming

		     Group Work: GUI Programming
		 (File $Date: 2020/06/30 18:05:44 $)

July 8, 2020

Work individually or in pairs to write a simple "timer" dialog in
Java. The dialog shall consist of three GUI components: a text field
to display the elapse time in the hh:mm:ss format and start/stop
buttons to start and stop the timer (see the lecture notes for
sample screenshots).

As suggested in the lecture notes, separate the model (TimerModel)
from the view and control (TimerDialog) to make the model reusable. In
fact, you will reuse the model class when you create an Android
version in the next lesson.

Code it incrementally.
  P1*: GUI consisting of a text field and two buttons.
      Swing features: JLabel, JButton, and layout managers such as
        BorderLayout and FlowLayout

  P2: Logic for starting and stopping the timer and displaying 
      the start/stop time.
      Swing features: Event, event handler and event registration
        ActionEvent, ActionListener, addActionListener
	long System.currentTimeMillies()

  P3: Display the elapsed time continuously, e.g., once per second
      Swing feature: single threadness
        To update the display periodically, use either 
	the javax.swing.Timer class:
          Timer(int, ActionListener), start(), stop()
        or the Thread class and SwingUtilities.invokeLater(Runnable)

  *Use the provided base code; you can introduce a subclass or 
   directly modify it.